Make listing directories a lot faster
Splitting sequences via head+tail appears to be really slow. I'm
not really sure where the slowdown comes from as the underlying
concrete data structure in this particular case is an array, but
it's better not to rely on that anyway. Converting the sequence to
a list upfront speeds things up a lot.