Support Unix sockets
Abusing NamedPipeServerStream for this purpose might not be the
best idea ever, though. Unfortunately net46 doesn't support the new
UnixDomainSocketEndPoint, so we can't really do it the proper way.
The protocol is called "netpipe" as it behaves a bit differently
from unix wrt non-absolute paths, doesn't even use Unix sockets on
Windows (not that we've checked how/if this thing works on Windows
anyway) and isn't really clearly described as Unix-socket-backed
on any implementation, suggesting it might not be something .NET
developers want to expose. Who knows, maybe they'll change it into
something else if they decide to break compatibility in some major
release. I don't even know if such implementation detail is even
covered by backwards compatibility guarantees, even more so if it's
not documented.