Target net45 instead
net45 versions of System.Collections.Immutable don't depend on
an external build of System.Memory so there's no conflict with
the Mono assembly. One thing I don't really understand is why
it's not possible to just use the version of System.Collections.Immutable
shipped with Mono, it's clearly there but it's not referenced
by default for some reason.